Get rid of MEMonitorIFluidHandler
