+++ src/main/java/appeng/parts/misc/FluidStorageBusPart.java
@@ -23,6 +23,7 @@ import java.util.Objects;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
+import appeng.me.storage.FluidHandlerAdapter;
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
@@ -80,7 +81,12 @@ public class FluidStorageBusPart extends AbstractStorageBusPart<IAEFluidStack>
         var handlerExtOpt = target
                 .getCapability(CapabilityFluidHandler.FLUID_HANDLER_CAPABILITY, targetSide);
         if (handlerExtOpt.isPresent()) {
-            return new FluidHandlerAdapter(handlerExtOpt.orElse(null), alertDevice);
+            return new FluidHandlerAdapter(handlerExtOpt.orElse(null)) {
+                @Override
+                protected void onInjectOrExtract() {
+                    alertDevice.run();
+                }
+            };
         }
 
         return null;
